int addd

+++ Array/10_Find_the_missing_number_in_a_array.py
@@ -1,3 +1,3 @@
 arr = [0,3,2,4,5,]
 n =len(arr)
-print(n*(n+1)/2  - sum(arr))
+print(int(n*(n+1)/2)  - sum(arr))
